conf vars no longer static for testing sake
Not super happy with this, but I wasn't super happy with using global vars in the first place. Making these not-static means other code could wreak havoc with this, but that would almost have to be intentional, and if anybody is doing that intentionally they deserve what they get.

Still, I should think about different ways of handling this.